Move enums to constants file
